requirements.txt
boto3==1.22.2
fastapi>=0.77.1
kafka==1.3.5
kafka-python==2.0.2
loguru==0.6.0
pre-commit==2.18.1
pydantic>=1.9.0
pytest>=7.1.2
python-dotenv==0.20.0
requests>=2.27.1
uvicorn>=0.17.6
